// src/interfaces/reward.interface.ts
export interface Reward {
  level: string;
  total: number;
  tradableGoldValue: number;
  goldValue: number;
  reward: {
    [key: string]: {
      min: number;
      max: number;
      avg: number;
    };
  };
}

export interface RewardsPresenterProps {
  categories: string[];
  category: string;
  onClickCategory: (category: string) => () => void;
  rewards: Reward[];
}